# fallback to default case when formatting month - duplicate default