It is safe to remove the switch case that returns default month formatting and use default instead.

```console
$ benchstat a.txt b.txt
goos: darwin
goarch: arm64
pkg: go.expect.digital/intl
cpu: Apple M2 Max
                         │    a.txt    │               b.txt                │
                         │   sec/op    │   sec/op     vs base               │
NewDateTime/fa-IR-12       207.8n ± 1%   203.3n ± 1%  -2.14% (p=0.000 n=10)
NewDateTime/lv-LV-12       277.8n ± 1%   254.9n ± 1%  -8.23% (p=0.000 n=10)
NewDateTime/dz-BT-12       280.5n ± 0%   256.1n ± 2%  -8.68% (p=0.000 n=10)
DateTime_Format/fa-IR-12   828.8n ± 0%   817.9n ± 1%  -1.31% (p=0.006 n=10)
DateTime_Format/lv-LV-12   188.3n ± 2%   188.5n ± 1%       ~ (p=0.987 n=10)
DateTime_Format/dz-BT-12   470.4n ± 1%   471.9n ± 1%       ~ (p=0.225 n=10)
geomean                    325.4n        314.4n       -3.40%

                         │   a.txt    │                b.txt                │
                         │    B/op    │    B/op     vs base                 │
NewDateTime/fa-IR-12       269.0 ± 0%   269.0 ± 0%       ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12       237.0 ± 0%   221.0 ± 0%  -6.75% (p=0.000 n=10)
NewDateTime/dz-BT-12       237.0 ± 0%   221.0 ± 0%  -6.75% (p=0.000 n=10)
DateTime_Format/fa-IR-12   104.0 ± 0%   104.0 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   24.00 ± 0%   24.00 ± 0%       ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   144.0 ± 0%   144.0 ± 0%       ~ (p=1.000 n=10) ¹
geomean                    132.6        129.5       -2.30%
¹ all samples are equal

                         │   a.txt    │                b.txt                 │
                         │ allocs/op  │ allocs/op   vs base                  │
NewDateTime/fa-IR-12       5.000 ± 0%   5.000 ± 0%        ~ (p=1.000 n=10) ¹
NewDateTime/lv-LV-12       6.000 ± 0%   5.000 ± 0%  -16.67% (p=0.000 n=10)
NewDateTime/dz-BT-12       6.000 ± 0%   5.000 ± 0%  -16.67% (p=0.000 n=10)
DateTime_Format/fa-IR-12   23.00 ± 0%   23.00 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/lv-LV-12   4.000 ± 0%   4.000 ± 0%        ~ (p=1.000 n=10) ¹
DateTime_Format/dz-BT-12   24.00 ± 0%   24.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean                    8.575        8.069        -5.90%
¹ all samples are equal
```